+++ Changelog.md
@@ -4,6 +4,7 @@ All significant changes are listed in this file.
 
 ## [1.1.5] - 2022-10-27
 
+- Add support for "series" movie metadata
 - Fixed incorrect count type showing on status bar
 
 ## [1.1.4] - 2022-10-26